+++ BrytecConfig/src/data/Config.h
@@ -29,7 +29,7 @@ public:
     std::string getName() { return m_path.stem().string(); }
 
     uint8_t getAssignedModuleAddress(std::shared_ptr<NodeGroup> nodeGroup);
-    uint8_t getAssignedPinAddress(std::shared_ptr<NodeGroup> nodeGroup);
+    uint16_t getAssignedPinAddress(std::shared_ptr<NodeGroup> nodeGroup);
     std::shared_ptr<Pin> getAssignedPin(std::shared_ptr<NodeGroup> nodeGroup);
     std::shared_ptr<Module> getAssignedModule(std::shared_ptr<NodeGroup> nodeGroup);
     bool getUsedOnBus(std::shared_ptr<NodeGroup> nodeGroup);
